requirements.txt
Pillow # For image processing > 'import PIL'
croniter # For parsing cron expressions > 'import croniter'